refactor: 타입 임포트 방식을 'type'으로 변경하고, 불필요한 변수 제거 및 오류 처리 개선
