+++ src/features/allergy/hooks/useAllergy.ts
@@ -21,6 +21,7 @@ export const useAllergy = () => {
 
 	useEffect(() => {
 		fetchAllergies()
+		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [])
 
 	return {
