+++ src/features/myInfo/components/BookmarkTab.tsx
@@ -1,10 +1,13 @@
 import React, { useState, useEffect } from 'react'
-import { RestaurantCard, Restaurant } from '@/features/myInfo/components/RestaurantCard'
-import { FilterButtons, FilterCategory } from '@/shared/components/FilterButtons'
-import { SortDropdown, SortOption } from '@/shared/components/SortDropdown'
+import { RestaurantCard } from '@/features/myInfo/components/RestaurantCard'
+import { FilterButtons } from '@/shared/components/FilterButtons'
+import { SortDropdown } from '@/shared/components/SortDropdown'
 import { Pagination } from '@/shared/components/Pagination'
 import { getBookmark } from '../api/bookmarkApi'
 import { getConvertCategory } from '@/shared/hooks/useCategory'
+import type { Restaurant } from '../types/recommendationHistory'
+import type { FilterCategory } from '@/shared/components/FilterButtons'
+import type { SortOption } from '@/shared/components/SortDropdown'
 
 export const BookmarkTab: React.FC = () => {
 	const [bookmarkedRestaurants, setBookmarkedRestaurants] = useState<Restaurant[]>([])
@@ -24,6 +27,7 @@ export const BookmarkTab: React.FC = () => {
 
 	useEffect(() => {
 		fetchBookmark()
+		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [selectedCategory, selectedSort, currentPage])
 
 	// 필터나 정렬 변경 시 첫 페이지로 리셋
